+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { TodosPage } from './pages/pages-logic';
+import { TodosPage } from './pages';
 import styles from './styles/App.module.scss';
 
 function App() {
