+++ src/components/card/card.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables';
+@use '../../styles/variables';
 
 .card {
   display: flex;
@@ -22,6 +22,7 @@
   gap: 10px;
   padding: 10px;
   font-size: 1.3rem;
+	overflow: hidden;
 }
 
 .buttons-wrapper-actions {
