+++ src/components/index.ts
@@ -1,7 +1,5 @@
-export { Button } from './ui/button/button';
-export { Input } from './ui/input/input';
-export { CardUI } from './ui/card/card';
-export { Tab } from './ui/tab/tab';
-export { Tabs } from './ui/tabs/tabs';
-
-export { Card } from './logic/card/card';
+export { Button } from './button/button';
+export { Input } from './input/input';
+export { Card } from './card/card';
+export { Tab } from './tab/tab';
+export { Tabs } from './tabs/tabs';
