+++ src/components/tab/tab.module.scss
@@ -1,0 +1,19 @@
+@use '../../styles/variables';
+
+.tab {
+  display: flex;
+  justify-content: center;
+  max-width: 150px;
+  width: 100%;
+  font-weight: bold;
+  font-size: 1.3rem;
+}
+
+.tab:hover {
+  color: variables.$primary-color-hover;
+}
+
+.tab-active {
+  color: variables.$primary-color;
+  border-bottom: 1px solid variables.$primary-color;
+}
