+++ src/components/tab/tab.tsx
@@ -1,4 +1,4 @@
-import { TStatusRU } from '../../../lib/types';
+import { TStatusRU } from '../tabs/status-types';
 import styles from './tab.module.scss';
 
 interface TTab {
