+++ src/components/tabs/tabs-constants.ts
@@ -1,3 +1,3 @@
-import { TStatusRU } from '../types';
+import { TStatusRU } from "./status-types";
 
 export const tabs: TStatusRU[] = ['Все', 'Сделано', 'В работе'];
