+++ src/components/tabs/tabs.module.scss
