--- src/components/ui/button/button.module.scss
@@ -1,32 +1,0 @@
-@use '../../../styles/variables';
-@use '../../../styles/mixins.scss';
-
-.button {
-  &-big {
-    @include mixins.button-common;
-    max-width: 150px;
-    width: 100%;
-    padding: 5px;
-  }
-
-  &-small {
-    @include mixins.button-common;
-    width: 40px;
-  }
-
-  &-red {
-    background-color: variables.$main-color;
-  }
-
-  &-red:hover {
-    background-color: variables.$main-color-hover;
-  }
-
-  &-blue {
-    background-color: variables.$secondary-color;
-  }
-
-  &-blue:hover {
-    background-color: variables.$secondary-color-hover;
-  }
-}
