--- src/components/ui/tab/tab.module.scss
@@ -1,19 +1,0 @@
-@use '../../../styles/variables';
-
-.tab {
-  display: flex;
-  justify-content: center;
-  max-width: 150px;
-  width: 100%;
-  font-weight: bold;
-  font-size: 1.3rem;
-}
-
-.tab:hover {
-  color: variables.$secondary-color-hover;
-}
-
-.tab-active {
-  color: variables.$secondary-color;
-  border-bottom: 1px solid variables.$secondary-color;
-}
