--- src/lib/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './button-constants';
-export * from './status-constants';
-export * from './tabs-constants';
