+++ src/lib/types/index.ts
@@ -1,2 +1,1 @@
 export * from './common-types';
-export * from './status-types';
