--- src/pages/pages-logic/index.ts
@@ -1,1 +1,0 @@
-export { TodosPage } from './todos-page';
