--- src/pages/pages-ui/index.ts
@@ -1,1 +1,0 @@
-export { TodosPageUI } from './todos-page-ui/todos-page';
