+++ src/styles/App.module.scss
@@ -6,5 +6,5 @@
   align-items: center;
   min-height: 100vh;
   padding: 20px;
-  background-color: variables.$main-background;
+  background-color: variables.$primary-background;
 }
