+++ src/styles/_variables.scss
@@ -1,7 +1,7 @@
-$main-color: rgb(255, 69, 58);
-$main-color-hover: rgb(255, 92, 81);
-$secondary-color: rgb(10, 132, 255);
-$secondary-color-hover: rgb(51, 153, 255);
-$main-background: rgb(220, 220, 220);
-$light-color: rgb(250, 250, 250);
-$light-color-hover: rgb(240, 240, 240);
+$primary-color: #0A84FF;
+$primary-color-hover: #3399FF;
+$secondary-color: #FF453A;
+$secondary-color-hover: #FF5C51;
+$primary-background: #DCDCDC;
+$light-color: #FAFAFA;
+$light-color-hover: #F0F0F0;
