+++ src/styles/index.css
@@ -61,6 +61,5 @@ video {
 
 body {
   min-height: 100vh;
-  line-height: 1.5;
   -webkit-font-smoothing: antialiased;
 }
