--- src/styles/mixins.scss
@@ -1,8 +1,0 @@
-@mixin button-common {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 40px;
-  border-radius: 5px;
-  color: white;
-}
